rename project to ENAos
add iso image in repository
minimize Makefile output
add shell function
fix stdio null functions
add more info to panic
many other refactoring changes

+++ src/include/string.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stddef.h>
+#include <stdint.h>
 #include <stdbool.h>
 
 extern char* strcpy(char* destination, const char* source);
@@ -9,5 +10,5 @@ extern bool strncmp(const char *a, const char *b, size_t length);
 extern char *strext(char *destination, char *source, char attribute);
 extern char *strtok(char *string, const char *separator);
 extern size_t strlen(const char *string);
-extern char *itoa(int value, char *string, int base);
+extern char *itoa(size_t value, char *string, int base);
 extern int atoi(const char *string);
